requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
packaging==25.0
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
spacy==3.7.2
